Refactor: UserService 통합 및 comment 반영 (#6)
