+++ src/main/java/EatPic/spring/domain/user/controller/UserController.java
@@ -1,12 +1,8 @@
 package EatPic.spring.domain.user.controller;
 
 import EatPic.spring.domain.user.dto.*;
-import EatPic.spring.domain.user.entity.User;
-import EatPic.spring.domain.user.service.UserCommandServiceImpl;
-import EatPic.spring.domain.user.service.UserQueryService;
 import EatPic.spring.domain.user.service.UserService;
 import EatPic.spring.global.common.ApiResponse;
-import EatPic.spring.global.common.BaseResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.servlet.http.HttpServletRequest;
@@ -22,32 +18,18 @@ import java.util.Map;
 @RestController
 @RequestMapping("/api/auth")
 @RequiredArgsConstructor
-public class UserController {
+public class UserController{
 
     private final UserService userService;
-    private final UserCommandServiceImpl userCommandService;
-    private final UserQueryService userQueryService;
 
     // 회원 가입 요청
     @PostMapping("/signup")
     @Operation(summary = "이메일 회원가입 요청")
     public ResponseEntity<SignupResponseDTO> signup(@Valid @RequestBody SignupRequestDTO request) {
-
-        User savedUser = userService.signup(request);
-
-        SignupResponseDTO response = SignupResponseDTO.builder()
-                .userId(savedUser.getId())
-                .email(savedUser.getEmail())
-                .nameId(savedUser.getNameId())
-                .nickname(savedUser.getNickname())
-                .marketingAgreed(savedUser.getMarketingAgreed())
-                .message("회원가입이 완료되었습니다.")
-                .build();
-
-        return ResponseEntity.ok(response);
+        return ResponseEntity.ok(userService.signup(request));
     }
 
-    // 필수 동의 약관 확인
+    // 필수 동의 약관 확인 -> feature/5 Refacotor 하기
     @RestControllerAdvice
     public static class GlobalExceptionHandler {
 
@@ -64,8 +46,8 @@ public class UserController {
 
     @PostMapping("/login/email")
     @Operation(summary = "이메일 로그인 요청")
-    public BaseResponse<LoginResponseDTO> login(@RequestBody @Valid LoginRequestDTO request) {
-        return BaseResponse.onSuccess(userCommandService.loginUser(request));
+    public ApiResponse<LoginResponseDTO> login(@RequestBody @Valid LoginRequestDTO request) {
+        return ApiResponse.onSuccess(userService.loginUser(request));
     }
 
     @GetMapping("/users")
@@ -73,7 +55,7 @@ public class UserController {
             security = { @SecurityRequirement(name = "JWT TOKEN") }
     )
     public ApiResponse<UserInfoDTO> getMyInfo(HttpServletRequest request) {
-        return ApiResponse.onSuccess(userQueryService.getUserInfo(request));
+        return ApiResponse.onSuccess(userService.getUserInfo(request));
     }
 }
