+++ src/main/java/EatPic/spring/domain/user/controller/UserRestController.java
@@ -1,7 +1,7 @@
 package EatPic.spring.domain.user.controller;
 
 import EatPic.spring.domain.user.dto.UserResponseDTO;
-import EatPic.spring.domain.user.service.UserCommandService;
+import EatPic.spring.domain.user.service.UserService;
 import EatPic.spring.global.common.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @RequestMapping("/api/community")
 public class UserRestController {
-    private final UserCommandService userCommandService;
+    private final UserService userService;
 
     @Operation(
             summary = "커뮤니티 상단 팔로잉 유저 아이콘",
@@ -23,7 +23,7 @@ public class UserRestController {
                                       @RequestParam(defaultValue = "1") int page,
                                       @RequestParam(defaultValue = "15") int size) {
         //todo : userId -> 본인
-        return ApiResponse.onSuccess(userCommandService.followingUserIconList(1L, page-1, size));
+        return ApiResponse.onSuccess(userService.followingUserIconList(1L, page-1, size));
     }
 
     @Operation(
@@ -31,6 +31,6 @@ public class UserRestController {
     @GetMapping("/me")
     @Tag(name = "User", description = "사용자 관련 API")
     public ApiResponse<UserResponseDTO.ProfileDto> myIcon() {
-        return ApiResponse.onSuccess(userCommandService.getMyIcon());
+        return ApiResponse.onSuccess(userService.getMyIcon());
     }
 }
