+++ src/main/java/EatPic/spring/domain/user/entity/User.java
@@ -67,5 +67,7 @@ public class User extends BaseEntity {
 
     @Column(name = "marketing_agreed", nullable = false)
     private Boolean marketingAgreed;
+
+    // notificationAgreed 추가 -> feature/5 Refacotor 하기
 }
