--- src/main/java/EatPic/spring/domain/user/exception/handler/UserHandler.java
@@ -1,10 +1,0 @@
-package EatPic.spring.domain.user.exception.handler;
-
-import EatPic.spring.global.common.code.BaseErrorCode;
-import EatPic.spring.global.common.exception.GeneralException;
-
-public class UserHandler extends GeneralException {
-    public UserHandler(BaseErrorCode errorCode) {
-        super(errorCode);
-    }
-}
