--- src/main/java/EatPic/spring/domain/user/service/UserQueryService.java
@@ -1,8 +1,0 @@
-package EatPic.spring.domain.user.service;
-
-import EatPic.spring.domain.user.dto.UserInfoDTO;
-import jakarta.servlet.http.HttpServletRequest;
-
-public interface UserQueryService {
-    UserInfoDTO getUserInfo(HttpServletRequest request);
-}
