+++ src/main/java/EatPic/spring/global/common/code/status/ErrorStatus.java
@@ -11,35 +11,43 @@ import org.springframework.http.HttpStatus;
 public enum ErrorStatus implements BaseErrorCode {
 
     // 기본 에러
-    _BAD_REQUEST(HttpStatus.BAD_REQUEST, "COMMON_400", "잘못된 요청입니다."),
-    _UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "COMMON_401", "인증이 필요합니다."),
-    _FORBIDDEN(HttpStatus.FORBIDDEN, "COMMON_403", "금지된 요청입니다."),
-    _INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "COMMON_500", "서버 에러. 관리자에게 문의 바랍니다."),
+    _BAD_REQUEST(false, HttpStatus.BAD_REQUEST, "COMMON_400", "잘못된 요청입니다."),
+    _UNAUTHORIZED(false, HttpStatus.UNAUTHORIZED, "COMMON_401", "인증이 필요합니다."),
+    _FORBIDDEN(false, HttpStatus.FORBIDDEN, "COMMON_403", "금지된 요청입니다."),
+    _INTERNAL_SERVER_ERROR(false, HttpStatus.INTERNAL_SERVER_ERROR, "COMMON_500", "서버 에러. 관리자에게 문의 바랍니다."),
 
-    MULTIPLE_FIELD_VALIDATION_ERROR(HttpStatus.BAD_REQUEST, "COMMON_001", "입력된 정보에 오류가 있습니다. 필드별 오류 메시지를 참조하세요."),
-    NO_MATCHING_ERROR_STATUS(HttpStatus.INTERNAL_SERVER_ERROR, "COMMON_002", "서버 에러. 일치하는 errorStatus를 찾을 수 없습니다."),
-    REQUEST_BODY_INVALID(HttpStatus.BAD_REQUEST, "COMMON_003", "요청 본문을 읽을 수 없습니다. 빈 문자열 또는 null이 있는지 확인해주세요."),
-    INVALID_DATE_FORMAT(HttpStatus.BAD_REQUEST, "COMMON_004", "요청한 날짜/시간 형식이 올바르지 않습니다. 형식을 확인해주세요."),
-    DUPLICATE_UNIQUE_KEY(HttpStatus.CONFLICT, "COMMON_005", "이미 처리된 요청입니다."),
+    MULTIPLE_FIELD_VALIDATION_ERROR(false, HttpStatus.BAD_REQUEST, "COMMON_001", "입력된 정보에 오류가 있습니다. 필드별 오류 메시지를 참조하세요."),
+    NO_MATCHING_ERROR_STATUS(false, HttpStatus.INTERNAL_SERVER_ERROR, "COMMON_002", "서버 에러. 일치하는 errorStatus를 찾을 수 없습니다."),
+    REQUEST_BODY_INVALID(false, HttpStatus.BAD_REQUEST, "COMMON_003", "요청 본문을 읽을 수 없습니다. 빈 문자열 또는 null이 있는지 확인해주세요."),
+    INVALID_DATE_FORMAT(false, HttpStatus.BAD_REQUEST, "COMMON_004", "요청한 날짜/시간 형식이 올바르지 않습니다. 형식을 확인해주세요."),
+    DUPLICATE_UNIQUE_KEY(false, HttpStatus.CONFLICT, "COMMON_005", "이미 처리된 요청입니다."),
 
     // Auth 관련 에러
-    AUTH_EXPIRED_TOKEN(HttpStatus.UNAUTHORIZED, "AUTH_001", "토큰이 만료되었습니다."),
-    AUTH_INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "AUTH_002", "토큰이 유효하지 않습니다."),
-    INVALID_USERNAME_OR_PASSWORD(HttpStatus.UNAUTHORIZED, "AUTH_003", "이메일 또는 비밀번호가 유효하지 않습니다."),
-    REQUEST_HEADER_INVALID(HttpStatus.BAD_REQUEST, "AUTH_004", "잘못된 요청 헤더입니다."),
-    AUTHENTICATION_FAILED(HttpStatus.UNAUTHORIZED, "AUTH_005", "유저 인증에 실패했습니다."),
-    UNAUTHORIZED_USER(HttpStatus.FORBIDDEN, "AUTH_006", "해당 요청에 대한 권한이 없는 유저입니다."),
-    LOGOUT_ACCESS_TOKEN(HttpStatus.UNAUTHORIZED, "AUTH_007", "로그아웃된 어세스 토큰 입니다"),
-    TEMP_TOKEN_NOT_ALLOWED(HttpStatus.BAD_REQUEST, "AUTH_014", "임시 토큰을 사용한 요청은 허용되지 않습니다."),
-    INTERNAL_AUTH_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "AUTH_009", "인증 처리 중 서버 에러가 발생했습니다."),
+    AUTH_EXPIRED_TOKEN(false, HttpStatus.UNAUTHORIZED, "AUTH_001", "토큰이 만료되었습니다."),
+    AUTH_INVALID_TOKEN(false, HttpStatus.UNAUTHORIZED, "AUTH_002", "토큰이 유효하지 않습니다."),
+    INVALID_USERNAME_OR_PASSWORD(false, HttpStatus.UNAUTHORIZED, "AUTH_003", "이메일 또는 비밀번호가 유효하지 않습니다."),
+    REQUEST_HEADER_INVALID(false, HttpStatus.BAD_REQUEST, "AUTH_004", "잘못된 요청 헤더입니다."),
+    AUTHENTICATION_FAILED(false, HttpStatus.UNAUTHORIZED, "AUTH_005", "유저 인증에 실패했습니다."),
+    UNAUTHORIZED_USER(false, HttpStatus.FORBIDDEN, "AUTH_006", "해당 요청에 대한 권한이 없는 유저입니다."),
+    LOGOUT_ACCESS_TOKEN(false, HttpStatus.UNAUTHORIZED, "AUTH_007", "로그아웃된 어세스 토큰 입니다"),
+    TEMP_TOKEN_NOT_ALLOWED(false, HttpStatus.BAD_REQUEST, "AUTH_014", "임시 토큰을 사용한 요청은 허용되지 않습니다."),
+    INTERNAL_AUTH_ERROR(false, HttpStatus.INTERNAL_SERVER_ERROR, "AUTH_009", "인증 처리 중 서버 에러가 발생했습니다."),
 
     // 카드 관련 응답
-    CARD_NOT_FOUND(HttpStatus.NOT_FOUND, "CARD_001", "해당 카드는 존재하지 않는 카드입니다."),
+    CARD_NOT_FOUND(false, HttpStatus.NOT_FOUND, "CARD_001", "해당 카드는 존재하지 않는 카드입니다."),
 
     // 댓글 관련 응답
-    COMMENT_NOT_FOUND(HttpStatus.NOT_FOUND, "COMMENT_001", "해당 댓글은 존재하지 않는 댓글입니다."),
-    CURSOR_NOT_FOUND(HttpStatus.BAD_REQUEST, "COMMENT_002", "유효하지 않은 커서입니다.");
+    COMMENT_NOT_FOUND(false, HttpStatus.NOT_FOUND, "COMMENT_001", "해당 댓글은 존재하지 않는 댓글입니다."),
+    CURSOR_NOT_FOUND(false, HttpStatus.BAD_REQUEST, "COMMENT_002", "유효하지 않은 커서입니다."),
 
+    // User Error
+    USER_NOT_FOUND(false, HttpStatus.NOT_FOUND, "USER_404", "존재하지 않는 사용자입니다."),
+    MEMBER_NOT_FOUND(false, HttpStatus.BAD_REQUEST, "USER_404", "사용자가 없습니다."),
+    INVALID_PASSWORD(false, HttpStatus.BAD_REQUEST, "USER_401", "비밀번호가 불일치합니다."),
+    INVALID_TOKEN(false, HttpStatus.BAD_REQUEST, "USER_403", "유효하지 않은 토큰입니다."),
+    DUPLICATE_JOIN_REQUEST(false, HttpStatus.BAD_REQUEST, "USER_400", "해당 이메일로 이미 가입된 사용자가 존재합니다.");
+
+    private final boolean isSuccess;
     private final HttpStatus httpStatus;
     private final String code;
     private final String message;
@@ -47,7 +55,8 @@ public enum ErrorStatus implements BaseErrorCode {
     @Override
     public ErrorReasonDTO getReason() {
         return ErrorReasonDTO.builder()
-                .isSuccess(false)
+                .isSuccess(isSuccess)
+                .httpStatus(httpStatus)
                 .code(code)
                 .message(message)
                 .build();
@@ -55,11 +64,7 @@ public enum ErrorStatus implements BaseErrorCode {
 
     @Override
     public ErrorReasonDTO getReasonHttpStatus() {
-        return ErrorReasonDTO.builder()
-                .httpStatus(httpStatus)
-                .isSuccess(false)
-                .code(code)
-                .message(message)
-                .build();
+        return getReason(); // 필요 시 httpStatus만 반환하도록 따로 구성 가능
     }
+
 }
